Update taxi_simulator.py
move print bill to date outside of menu_choice tree
change distance input from int to float

+++ prac_08/taxi_simulator.py
@@ -14,9 +14,9 @@ while menu_choice != 'q':
         taxi_choice = int(input("Choose taxi: "))
         current_taxi = taxis[taxi_choice]
     if menu_choice == 'd' and current_taxi is not None:
-        distance = int(input("Drive how far? "))
+        distance = float(input("Drive how far? "))
         current_taxi.drive(distance)
         print("Your {} trip cost you ${:.2f}".format(current_taxi.name, current_taxi.get_fare()))
         total_bill += current_taxi.get_fare()
-        print("Bill to date: ${:.2f}".format(total_bill))
+    print("Bill to date: ${:.2f}".format(total_bill))
     menu_choice = input("q)uit, c)hoose taxi, d)rive \n>>> ").lower()
